Moved debugger into own package
