+++ src/main/java/com/ccs/baby/debug/Debugger.java
@@ -1,4 +1,4 @@
-package com.ccs.baby.ui;
+package com.ccs.baby.debug;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
